refactor: move entity validation logic to use cases

+++ EventWebApp.Application/UseCases/Event/DeleteEventUseCase.cs
@@ -1,19 +1,28 @@
-﻿using EventWebApp.Core.Interfaces;
+﻿using EventWebApp.Application.Exceptions;
+using EventWebApp.Core.Interfaces;
 
 namespace EventWebApp.Application.UseCases.Event
 {
-    public class DeleteEventUseCase
+  public class DeleteEventUseCase
+  {
+    private readonly IEventRepository eventRepository;
+
+    public DeleteEventUseCase(IEventRepository eventRepository)
     {
-        private readonly IEventRepository eventRepository;
-
-        public DeleteEventUseCase(IEventRepository eventRepository)
-        {
-            this.eventRepository = eventRepository;
-        }
-
-        public async Task ExecuteAsync(Guid id)
-        {
-            await eventRepository.DeleteAsync(id);
-        }
+      this.eventRepository = eventRepository;
     }
+
+    public async Task ExecuteAsync(Guid id)
+    {
+      // Проверка существования события
+      var existingEvent = await eventRepository.GetByIdAsync(id);
+      if (existingEvent == null)
+      {
+        throw new NotFoundException("Event not found", ErrorCodes.EventNotFound);
+      }
+
+      // Удаление готовой сущности
+      await eventRepository.DeleteAsync(id);
+    }
+  }
 }
